Assignment 3 - Seshat & Cooc Mx/requirements.txt
pyenchant
selenium
pdfminer
wordninja
